[FEAT] app lecture api 추가
