+++ operation-api/src/main/java/org/sopt/makers/operation/app/lecture/dto/response/LectureResponseType.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.operation.service.app.lecture.dto.response;
+package org.sopt.makers.operation.app.lecture.dto.response;
 
 public enum LectureResponseType {
     NO_SESSION, HAS_ATTENDANCE, NO_ATTENDANCE
